pyproject.toml
[project]
name = "memory-examples"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain-neo4j>=0.4.0",
    "langchain-openai>=0.3.18",
    "langgraph>=0.4.7",
    "mem0ai>=0.1.102",
    "python-dotenv>=1.1.0",
    "rank-bm25>=0.2.2",
]
